==> week2/02_scripts/02.exploration_empty.Rmd <==
---
title: "RNA-seq course - week 2. Exploration"
author: "Oleksandr Petrenko"
date: "`r Sys.Date()`"
output:
  pdf_document:
    toc: yes
  html_document:
    code_folding: hide
    df_print: paged
    highlights: pygments
    number_sections: yes
    self_contained: yes
    theme: default
    toc: yes
    toc_float:
      collapsed: yes
      smooth_scroll: yes
editor_options: 
  chunk_output_type: inline
---

## Setup

```{r setup}
knitr::opts_knit$set(root.dir = '/home/opetrenko/projects/2023_rnaseq_course/week2')
setwd("/home/opetrenko/projects/2023_rnaseq_course/week2")
```

```{r install-libraries}
#if (!require("BiocManager", quietly = TRUE))
#    install.packages("BiocManager")

#BiocManager::install("PCAtools")
#BiocManager::install("DESeq2")
#install.packages("pheatmap")
```

```{r load-libraries-and-files}
library(PCAtools)
#library(pheatmap)
library(DESeq2)
library(tidyverse)

getwd()
count_matrix <- read.csv("01_datasets/counts.csv", row.names = 1)
count_matrix_filtered <- read.csv("03_outputs/01_count_matrix_filtered.csv", row.names = 1)
metadata_table <- read.csv("01_datasets/metadata.csv", row.names = 1)
```
## PCA

```{r pca-raw}
pca_calculated <- pca(count_matrix, metadata = metadata_table, removeVar = 0.1)

biplot(pca_calculated, colby = "fibroblast_line", legendPosition = "right", lab = NULL)


biplot(pca_calculated, colby = "treatment", legendPosition = "right", lab = NULL)
metadata_table

# eigenvector
# eigenvalue
```

```{r normalization}
dds <- DESeqDataSetFromMatrix(countData = count_matrix_filtered, colData = metadata_table,
                              design = ~1)


vsd <- vst(dds, blind = TRUE)

counts_vst_filtered <- assay(vsd) %>% as.data.frame()

counts_vst_filtered
```

```{r pca-normalized}
pca_calculated <- pca(counts_vst_filtered, metadata = metadata_table, removeVar = 0.1)

biplot(pca_calculated, colby = "fibroblast_line", legendPosition = "right", lab = NULL)

biplot(pca_calculated, colby = "treatment", legendPosition = "right", lab = NULL)
metadata_table
```

```{r pc-exploration}
pairsplot(pca_calculated, colby = "treatment")
```

```{r}
plotloadings(pca_calculated, components = getComponents(pca_calculated,
             c(1, 2, 5)), rangeRetain = 0.1, drawConnectors = TRUE)
```


## Hierarchical clustering

```{r}
matrix_correlations <- cor(counts_vst_filtered)

f1 <- pheatmap(matrix_correlations, annotation = metadata_table, show_rownames = TRUE,
         show_colnames = FALSE)
```

## Outputs

```{r}
ggsave("03_outputs/02_PCA.png", F1, device = "png")
ggsave("03_outputs/02_clustering.png", f1, device = "png")
ggsave()
```
