==> week2/02_scripts/04_normalization_comparison.Rmd <==
---
title: "RNA-seq course - week 2. Normalization comparison"
author: "Oleksandr Petrenko"
date: "`r Sys.Date()`"
output:
  pdf_document:
    toc: yes
  html_document:
    code_folding: hide
    df_print: paged
    highlights: pygments
    number_sections: yes
    self_contained: yes
    theme: default
    toc: yes
    toc_float:
      collapsed: yes
      smooth_scroll: yes
editor_options: 
  chunk_output_type: inline
---

## Setup

```{r setup}
knitr::opts_knit$set(root.dir = '~/projects/2023_rnaseq_course/week2')
setwd("~/projects/2023_rnaseq_course/week2")
```


```{r install-packages}
#BiocManager::install("DEGreport")
```


```{r load-libraries-and-files}
library(DESeq2)
library(tidyverse)

getwd()
count_matrix_filtered <- read.csv("03_outputs/01_count_matrix_filtered.csv", row.names = 1)
```

```{r}
count_matrix_filtered
```


## VST: what does it do?

```{r}
fibroblast_genes <- c("ENSG00000108821", # COL1A1
                      "ENSG00000164692", # COL1A2
                      "ENSG00000168542", # COL3A1
                      "ENSG00000149591", # TAGLN
                      "ENSG00000107796", # ACTA2
                      "ENSG00000198467", # TPM2
                      
                      "ENSG00000001497", # LAS1L
                      "ENSG00000000460", # C1orf112
                      "ENSG00000001461" # NIPAL3
                      ) 
```

```{r}
count_matrix_filtered[fibroblast_genes,]
```

```{r}
a <- ggplot(count_matrix_filtered) + geom_histogram(aes(x = GSM6214563), stat = "bin", bins = 200) +
        xlab("Raw expression counts") + 
        ylab("Number of genes")
  
```
```{r}
counts_mean_var <- count_matrix_filtered %>% 
  mutate(row = row_number()) %>%
  tidyr::pivot_longer(-row) %>%
  group_by(row) %>%
  summarize(mean = mean(value),
            var = var(value)) %>%
  bind_cols(count_matrix_filtered, .) %>%
  select(mean, var) %>%
  tibble::rownames_to_column("gene")
```

```{r}
ggplot(counts_mean_var) + geom_point(aes(x = mean, y = var)) +
                          scale_y_log10() +
                          scale_x_log10() +
                          geom_abline(intercept = 0, slope = 1, color = "red")
```

```{r}
counts_vst <- varianceStabilizingTransformation(count_matrix_filtered %>% as.matrix())
```

```{r}
counts_vst[fibroblast_genes, ] %>% as.data.frame()
```

```{r}
b <- ggplot(counts_vst %>% as.data.frame() %>% filter(GSM6214563 > 5)) + 
        geom_histogram(aes(x = GSM6214563), stat = "bin", bins = 200) +
        xlab("Raw expression counts") + 
        ylab("Number of genes")
  
```

```{r}
a
b
```

```{r}
vst_mean_var <- counts_vst %>% as.data.frame() %>%
  mutate(row = row_number()) %>%
  tidyr::pivot_longer(-row) %>%
  group_by(row) %>%
  summarize(mean = mean(value),
            var = var(value)) %>%
  bind_cols(counts_vst, .) %>%
  select(mean, var) %>%
  tibble::rownames_to_column("gene")
```


```{r}
ggplot(vst_mean_var) + geom_point(aes(x = mean, y = var)) +
                          scale_y_log10() +
                          scale_x_log10() +
                          geom_abline(intercept = 0, slope = 1, color = "red")
```
